// serve/src/dashboard/dto/symbol.dto.ts
export type SymbolProps = {
  symbol: string;
  name: string;
  type: string;
  region: string;
  marketOpen: string;
  marketClose: string;
  timezone: string;
  currency: string;
  matchScore: string;
};

export class SymbolDTO {
  constructor(props: SymbolProps) {
    Object.assign(this, props);
  }

  symbol: string;
  name: string;
  type: string;
  region: string;
  marketOpen: string;
  marketClose: string;
  timezone: string;
  currency: string;
  matchScore: string;
}